respond to pr comments
